+++ system/src/Grav/Console/GravCommand.php
@@ -10,7 +10,6 @@
 namespace Grav\Console;
 
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Exception\LogicException;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -21,19 +20,6 @@ use Symfony\Component\Console\Output\OutputInterface;
 class GravCommand extends Command
 {
     use ConsoleTrait;
-
-    /**
-     * @param string|null $name The name of the command; passing null means it must be set in configure()
-     *
-     * @throws LogicException When the command name is empty
-     */
-    public function __construct(string $name = null)
-    {
-        parent::__construct($name);
-
-        // Add --env option.
-        $this->addEnvOption();
-    }
 
     /**
      * @param InputInterface  $input